move enum to pickup_state
